# Use a real reporter in handler benchmark to surface allocations.